fix scenes

+++ Assets/Script/Model/Weapon/Sword.cs
@@ -5,5 +5,5 @@
     public float speed { get => -1f; }
     public float defence { get => 3f; }
     public float force { get => 3f; }
-    public float attackSpeed { get => -1f; }
+    public int attackSpeed { get => -1; }
 }
